:recycle: Refactor useReducerFilter into custom hook
