+++ src/App.tsx
@@ -4,6 +4,7 @@ import styles from "./styles.module.scss";
 import SearchBar from "components/SearchBar";
 import Filter from "components/Filter";
 import { filterPokemon } from "./utils";
+import usePokemonFilterState from "components/Filter/usePokemonFilterState";
 
 export type BaseStats = {
   hp: number;
@@ -24,48 +25,11 @@ export type Pokemon = {
   baseStats: BaseStats;
 };
 
-export type FilterRange = [number, number];
-
-export type ValueFilters = {
-  weight: FilterRange;
-};
-
-export enum FilterAction {
-  SetMaxWeight = "setMaxWeight",
-  SetMinWeight = "setMinWeight",
-  Reset = "reset",
-}
-
-export type ActionType = {
-  type: FilterAction;
-  payload: number;
-};
-
-const MIN_FILTER_VALUE = 0;
-const MAX_FILTER_VALUE = 9999;
-
-const init = (): ValueFilters => {
-  return { weight: [MIN_FILTER_VALUE, MAX_FILTER_VALUE] };
-};
-
-const reducer = (state: ValueFilters, action: ActionType): ValueFilters => {
-  switch (action.type) {
-    case FilterAction.SetMaxWeight:
-      return { ...state, weight: [state.weight[0], action.payload] };
-    case FilterAction.SetMinWeight:
-      return { ...state, weight: [action.payload, state.weight[1]] };
-    case FilterAction.Reset:
-      return init();
-    default:
-      throw new Error();
-  }
-};
-
 const App = (): JSX.Element => {
   const [pokemons, setPokemons] = useState<Array<Pokemon>>([]);
   const [filteredPokemon, setFilteredPokemon] = useState<Array<Pokemon>>([]);
   const [searchTerm, setSearchTerm] = useState("");
-  const [state, dispatch] = useReducer(reducer, init());
+  const [state, dispatch] = usePokemonFilterState();
 
   const url = process.env.POKEMON_LIST;
